move home meter validator in separate package
