+++ d3a_interface/validators/utils.py
@@ -1,0 +1,21 @@
+"""
+Copyright 2018 Grid Singularity
+This file is part of D3A.
+
+This program is free software: you can redistribute it and/or modify it under the terms of the
+GNU General Public License as published by the Free Software Foundation, either version 3 of the
+License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along with this program. If
+not, see <http://www.gnu.org/licenses/>.
+"""
+from d3a_interface.exceptions import D3ADeviceException
+
+# TODO: put this in the base class? Better: create a RangeLimitValidator class
+def validate_range_limit(initial_limit, value, final_limit, error_message):
+    if not initial_limit <= value <= final_limit:
+        raise D3ADeviceException(error_message)
